feat: 캘린더 뷰모델 공통 기능 추상화, 타입 별로 분리(#39)
